fix: Separate unit and integration test phase with Maven failsafe plugin
Should stabilize the CI build where unit tests and integration tests conflicts happen occasionally
