+++ java/steps/yaks-jms/src/test/java/org/citrusframework/yaks/jms/JmsFeature_IT.java
@@ -15,19 +15,16 @@
  * limitations under the License.
  */
 
-package org.citrusframework.yaks.camel;
+package org.citrusframework.yaks.jms;
 
 import io.cucumber.junit.Cucumber;
 import io.cucumber.junit.CucumberOptions;
 import org.junit.runner.RunWith;
 
-/**
- * @author Christoph Deppisch
- */
 @RunWith(Cucumber.class)
 @CucumberOptions(
         extraGlue = { "org.citrusframework.yaks.standard" },
         plugin = { "pretty", "com.consol.citrus.cucumber.CitrusReporter" }
 )
-public class CamelFeatureTest {
+public class JmsFeature_IT {
 }
